Small fixes in anticipation of merge
* Improved perf info table: fixed arrow thread count, queue overhead
* Disabled JQueueWithBarriers, boosting Arrow performance
* Set sensible chunksizes in JTopologyBuilder, possibly improving perf someday
* Minor tweaks so that everything compiles again on Linux (weirdly, it compiled on Mac just fine)

+++ src/lib/JANA/JArrow.h
@@ -67,7 +67,7 @@ public:
         return _chunksize;
     }
 
-    void update_thread_count(size_t thread_count_delta) {
+    void update_thread_count(int thread_count_delta) {
         std::lock_guard<std::mutex> lock(_mutex);
         _thread_count += thread_count_delta;
     }
